refactor client cmds
